Better naming convention for interface, move errors to interface level, add info to README.md
